# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Recipes API\nThis is a sample recipes API",
    "contact": {
      "name": "Siddhartha Sitaula",
      "url": "https://sitaulasiddhartha2002@gmail.com",
      "email": "sitaulasiddhartha2002@gmail.com"
    },
    "version": "1.0.0"
  },
  "host": "localhost:8080",
  "basePath": "/",
  "paths": {
    "/recipes": {
      "get": {
        "description": "Returns list of recipes",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "ListRecipes"
      },
      "post": {
        "description": "Request body of the POST /recipes",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "NewRecipes",
        "parameters": [
          {
            "x-go-name": "ID",
            "name": "id",
            "in": "query"
          },
          {
            "type": "string",
            "x-go-name": "Name",
            "name": "name",
            "in": "query"
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            },
            "x-go-name": "Tags",
            "name": "tags",
            "in": "query"
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            },
            "x-go-name": "Ingredients",
            "name": "ingredients",
            "in": "query"
          },
          {
            "type": "array",
            "items": {
              "type": "string"
            },
            "x-go-name": "Instructions",
            "name": "instructions",
            "in": "query"
          },
          {
            "type": "string",
            "format": "date-time",
            "x-go-name": "PublishedAt",
            "name": "publishedAt",
            "in": "query"
          }
        ]
      }
    },
    "/recipes/{id}": {
      "put": {
        "description": "Update an existing recipe",
        "produces": [
          "application/json"
        ],
        "tags": [
          "recipes"
        ],
        "operationId": "updateRecipe",
        "parameters": [
          {
            "type": "string",
            "description": "ID of the recipe",
            "name": "id",
            "in": "path",
            "required": true
          }
        ]
      }
    },
    "/refresh": {
      "post": {
        "description": "Get new token in exchange for an old one",
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "operationId": "refresh"
      }
    }
  }
}